Fixes to Payout Samples
- Refactored Runner to enable sample chaining in command line
