+++ samples/app.rb
@@ -1,19 +1,9 @@
 # Configure Bundler
 require 'bundler/setup'
+require './runner.rb'
+
 Bundler.require :default, :sample
 
-# Run Samples on different scope
-module RunSample
-  def self.logger
-    PayPal::SDK::Core::Config.logger
-  end
-
-  def self.run(file, variable)
-    object_binding = binding
-    object_binding.eval(File.read("./#{file}"))
-    object_binding.eval(variable)
-  end
-end
 
 class App < Sinatra::Application
   enable :sessions
