+++ samples/payouts/cancel.rb
@@ -1,4 +1,6 @@
 require 'paypal-sdk-rest'
+require './runner.rb'
+
 include PayPal::SDK::REST
 include PayPal::SDK::Core::Logging
