+++ samples/payouts/create.rb
@@ -1,5 +1,6 @@
 require 'paypal-sdk-rest'
 require 'securerandom'
+require './runner.rb'
 
 include PayPal::SDK::REST
 include PayPal::SDK::Core::Logging
