+++ samples/payouts/get_item_status.rb
@@ -1,4 +1,5 @@
 require 'paypal-sdk-rest'
+require './runner.rb'
 include PayPal::SDK::REST
 include PayPal::SDK::Core::Logging
